Task_5_(4th,Feb)/backend/requirement.txt
fastapi
uvicorn
python-dotenv
authlib
pyjwt
requests
httpx
itsdangerous
